+++ examples/parse_from_iface.rs
@@ -8,6 +8,7 @@ use mutnet::ipv4::Ipv4Methods;
 use mutnet::ipv6::Ipv6Methods;
 use mutnet::multi_step_parser::{parse_network_data, EthernetMultiStepParserResult};
 use mutnet::tcp::TcpMethods;
+use mutnet::udp::UdpMethods;
 use std::io::{stdout, Write};
 
 #[derive(Copy, Clone, Debug, Default)]
@@ -21,9 +22,12 @@ struct Data {
     ipv4_dst: Option<Ipv4Address>,
     ipv6_next_hdr: Option<InternetProtocolNumber>,
     ipv6_src: Option<Ipv6Addr>,
+    fragment: Option<bool>,
     eth_ether_type: Option<EtherType>,
     tcp_source_port: Option<u16>,
     tcp_destination_port: Option<u16>,
+    udp_source_port: Option<u16>,
+    udp_destination_port: Option<u16>,
 }
 
 fn main() {
@@ -44,7 +48,7 @@ fn main() {
 
         match cap.next_packet() {
             Ok(packet) => {
-                match parse_network_data::<_, 10>(packet.data, 0, true, true) {
+                match parse_network_data::<_, 10>(packet.data, 0, true, true, true) {
                     Ok(EthernetMultiStepParserResult::Ethernet(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                     }
@@ -63,19 +67,39 @@ fn main() {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv4(&data_buffer, &mut output);
                     }
+                    Ok(EthernetMultiStepParserResult::FragmentIpv4Eth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv4(&data_buffer, &mut output);
+                        log_fragment(&mut output);
+                    }
                     Ok(EthernetMultiStepParserResult::TcpIpv4Eth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv4(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
                     }
+                    Ok(EthernetMultiStepParserResult::UdpIpv4Eth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv4(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
+                    }
                     Ok(EthernetMultiStepParserResult::Ipv4VlanEth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv4(&data_buffer, &mut output);
+                    }
+                    Ok(EthernetMultiStepParserResult::FragmentIpv4VlanEth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv4(&data_buffer, &mut output);
+                        log_fragment(&mut output);
                     }
                     Ok(EthernetMultiStepParserResult::TcpIpv4VlanEth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv4(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
+                    }
+                    Ok(EthernetMultiStepParserResult::UdpIpv4VlanEth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv4(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
                     }
                     Ok(EthernetMultiStepParserResult::Ipv6Eth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
@@ -86,6 +110,11 @@ fn main() {
                         log_ipv6(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
                     }
+                    Ok(EthernetMultiStepParserResult::UdpIpv6Eth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv6(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
+                    }
                     Ok(EthernetMultiStepParserResult::Ipv6VlanEth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv6(&data_buffer, &mut output);
@@ -94,6 +123,11 @@ fn main() {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv6(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
+                    }
+                    Ok(EthernetMultiStepParserResult::UdpIpv6VlanEth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv6(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
                     }
                     Ok(EthernetMultiStepParserResult::Ipv6ExtsIpv6Eth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
@@ -104,6 +138,11 @@ fn main() {
                         log_ipv6(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
                     }
+                    Ok(EthernetMultiStepParserResult::UdpIpv6ExtsIpv6Eth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv6(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
+                    }
                     Ok(EthernetMultiStepParserResult::Ipv6ExtsIpv6VlanEth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv6(&data_buffer, &mut output);
@@ -113,13 +152,20 @@ fn main() {
                         log_ipv6(&data_buffer, &mut output);
                         log_tcp(&data_buffer, &mut output);
                     }
+                    Ok(EthernetMultiStepParserResult::UdpIpv6ExtsIpv6VlanEth(data_buffer)) => {
+                        log_ethernet(&data_buffer, &mut output);
+                        log_ipv6(&data_buffer, &mut output);
+                        log_udp(&data_buffer, &mut output);
+                    }
                     Ok(EthernetMultiStepParserResult::FragmentIpv6ExtsIpv6Eth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv6(&data_buffer, &mut output);
+                        log_fragment(&mut output);
                     }
                     Ok(EthernetMultiStepParserResult::FragmentIpv6ExtsIpv6VlanEth(data_buffer)) => {
                         log_ethernet(&data_buffer, &mut output);
                         log_ipv6(&data_buffer, &mut output);
+                        log_fragment(&mut output);
                     }
                     // Checksum errors are normal for outgoing packets if your system is using
                     // hardware offload.
@@ -136,6 +182,9 @@ fn main() {
         }
         if let Some(arp_sender_prot_addr) = output.arp_sender_prot_addr {
             writeln!(lock, "Arp sender prot addr: {arp_sender_prot_addr:?}").unwrap();
+        }
+        if let Some(eth_ether_type) = output.eth_ether_type {
+            writeln!(lock, "Ethertype {eth_ether_type:?}").unwrap();
         }
         if let Some(ipv4_prot) = output.ipv4_prot {
             writeln!(lock, "IPv4 prototocol: {ipv4_prot:?}").unwrap();
@@ -165,8 +214,8 @@ fn main() {
         if let Some(ipv6_src) = output.ipv6_src {
             writeln!(lock, "IPv6 SRC {ipv6_src:X?}").unwrap();
         }
-        if let Some(eth_ether_type) = output.eth_ether_type {
-            writeln!(lock, "Ethertype {eth_ether_type:?}").unwrap();
+        if let Some(is_fragment) = output.fragment {
+            writeln!(lock, "Is fragment: {is_fragment:?}").unwrap();
         }
         if let Some(tcp_source_port) = output.tcp_source_port {
             writeln!(lock, "TCP src port: {tcp_source_port:?}").unwrap();
@@ -174,7 +223,14 @@ fn main() {
         if let Some(tcp_destination_port) = output.tcp_destination_port {
             writeln!(lock, "TCP dst port: {tcp_destination_port:?}").unwrap();
         }
+        if let Some(udp_source_port) = output.udp_source_port {
+            writeln!(lock, "UDP src port: {udp_source_port:?}").unwrap();
+        }
+        if let Some(udp_destination_port) = output.udp_destination_port {
+            writeln!(lock, "UDP dst port: {udp_destination_port:?}").unwrap();
+        }
         writeln!(lock, "Dropped packets {}", output.dropped_packets).unwrap();
+        writeln!(lock, "------------------").unwrap();
     }
 }
 
@@ -194,6 +250,10 @@ fn log_ipv4(packet: &impl Ipv4Methods, data: &mut Data) {
     data.ipv4_dst = Some(packet.ipv4_destination());
 }
 
+fn log_fragment(data: &mut Data) {
+    data.fragment = Some(true);
+}
+
 fn log_ipv6(packet: &impl Ipv6Methods, data: &mut Data) {
     data.ipv6_next_hdr = Some(packet.ipv6_typed_next_header().unwrap());
     data.ipv6_src = Some(packet.ipv6_source());
@@ -202,4 +262,9 @@ fn log_ipv6(packet: &impl Ipv6Methods, data: &mut Data) {
 fn log_tcp(packet: &impl TcpMethods, data: &mut Data) {
     data.tcp_source_port = Some(packet.tcp_source_port());
     data.tcp_destination_port = Some(packet.tcp_destination_port());
+}
+
+fn log_udp(packet: &impl UdpMethods, data: &mut Data) {
+    data.udp_source_port = Some(packet.udp_source_port());
+    data.udp_destination_port = Some(packet.udp_destination_port());
 }