+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# Unreleased
+# 0.3.0 (24. October 2023)
 ### Added
-- UDP
+- **breaking** UDP support for `parse_network_data()`
+- **breaking** IPv4 fragmentation support for `parse_network_data()`
+- UDP and IPv4 fragment support in parse_from_iface example
+
+# 0.2.0 (24. October 2023)
+### Added
+- **breaking** UDP
 - CI for tests
 
 ### Changed
@@ -10,6 +16,6 @@
 - TCP docs ([62cb561](https://github.com/arctic-alpaca/mutnet/commit/62cb5614b819d304bd38c3239d34c3cfff07f500) & [9601e0b](https://github.com/arctic-alpaca/mutnet/commit/9601e0bd56c1ee3b023453573c4fa830362b1ea3))
 - TCP/UDP checksum now uses information from the IPv4/IPv6 header in the pseudo header ([eb14bf4](https://github.com/arctic-alpaca/mutnet/commit/eb14bf42dfdec29e10b9bcd520bbd8075282ed83))
 
-# 0.1.0 (19. September, 2023)
+# 0.1.0 (19. September 2023)
 ### Added
 - Initial release